Allow admin to edit Member without supplying user's password
See:
https://github.com/sferik/rails_admin/issues/2265#issuecomment-102791278

Prior to this change, verifying a member failed due to password
constraints on the underlying model.

This change adds rails_admin  configuration to prevent password fields
from being considered for editing. Additionally, this changes the admin
console to allow an admin to reset a member’s password via the
administrative interface.